refactor: moved time validation to service
